aider: Remove the "Create Project" button.
# Aider chat conversation:

USER: Remove the create project `button`

ASSISTANT: To remove the "Create Project" button, we need to remove the corresponding code from the `10xdev/src/NewWelcome.js` file.

Here is the *edit block* for the change:

```javascript
10xdev/src/NewWelcome.js
<<<<<<< HEAD
            <button
            onClick={handleCreateClick}
            className="bg-blue-900 hover:bg-blue-800 font-bold px-6 py-2 my-6 text-white px-2 rounded"
            >
              Create Project
            </button>
=======
>>>>>>> updated
```

This *edit block* removes the button that triggers the `handleCreateClick` function. If there are other buttons or elements you want to remove or modify, please specify them.