aider: REMOVE the handleCreateClick function.
# Aider chat conversation:

USER: REMOVE THE handleCreateClick function

ASSISTANT: To remove the `handleCreateClick` function, we need to remove the corresponding code from the `10xdev/src/NewWelcome.js` file.

Here is the *edit block* for the change:

```javascript
10xdev/src/NewWelcome.js
<<<<<<< HEAD
  const handleCreateClick = (e) => {
      e.preventDefault();
      navigate("/clarify");
      }
=======
>>>>>>> updated
```

This *edit block* removes the `handleCreateClick` function. If there are other functions or elements you want to remove or modify, please specify them.